infotainment: add WatchServerResponses to get streaming session responses
After creating a streaming session, responses can be received from
the infotainment domain. This adds a function to get a channel for
receiving those messages. As a hack the last request ID is kept
temporarily in the dispatcher as for receiving messages the request
ID for the session creation is used in the stream responses.